refactor css for column header
